Imported using git-ubuntu import.

Publish parent: ce5ae0d07a3c55822a76606fcd327ffa7a1f94d9
Changelog parent: 7695d33edb09a754ac37c85e57b9aebe4d7f331f
Unapplied parent: 8ba6299627bf81d255d5e89a84845338449de8b0

+++ time/clocktest.c
@@ -28,9 +28,9 @@ main (int argc, char ** argv)
   while (!gotit);
   stop = clock ();
 
-  printf ("%jd clock ticks per second (start=%jd,stop=%jd)\n",
-	  (intmax_t) (stop - start), (intmax_t) start, (intmax_t) stop);
-  printf ("CLOCKS_PER_SEC=%jd, sysconf(_SC_CLK_TCK)=%ld\n",
-	  (intmax_t) CLOCKS_PER_SEC, sysconf(_SC_CLK_TCK));
+  printf ("%ld clock ticks per second (start=%ld,stop=%ld)\n",
+	  stop - start, start, stop);
+  printf ("CLOCKS_PER_SEC=%ld, sysconf(_SC_CLK_TCK)=%ld\n",
+	  CLOCKS_PER_SEC, sysconf(_SC_CLK_TCK));
   return 0;
 }
